Clean up a tad, add helper for logged in user lookup
